requirements.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
google-generativeai>=0.7.0
python-dotenv
